organize reusable meta/composite instances
