+++ modules/db/src/main/scala/gem/dao/meta/Level.scala
@@ -1,0 +1,16 @@
+// Copyright (c) 2016-2017 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package gem.dao.meta
+
+import doobie._
+import java.util.logging.Level
+
+trait LevelMeta {
+
+  // Java Log Levels (not nullable)
+  implicit val levelMeta: Meta[Level] =
+    Meta[String].xmap(Level.parse, _.getName)
+
+}
+object LevelMeta extends LevelMeta
