+++ modules/db/src/main/scala/gem/dao/meta/ObservationId.scala
@@ -1,0 +1,16 @@
+// Copyright (c) 2016-2017 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package gem.dao.meta
+
+import doobie._
+import gem.Observation
+
+trait ObservationIdMeta {
+
+  // Observation.Id as string
+  implicit val ObservationIdMeta: Meta[Observation.Id] =
+    Meta[String].xmap(Observation.Id.unsafeFromString, _.format)
+
+}
+object ObservationIdMeta extends ObservationIdMeta
